+++ _wotlk/Interface/Addons/Myslot/Myslot.lua
@@ -102,7 +102,7 @@ end
 function MySlot:GetMacroInfo(macroId)
     -- {macroId ,icon high 8, icon low 8 , namelen, ..., bodylen, ...}
 
-    local name, iconTexture, body, isLocal = GetMacroInfo(macroId)
+    local name, iconTexture, body  = GetMacroInfo(macroId)
 
     if not name then
         return nil
@@ -159,6 +159,28 @@ function MySlot:GetActionInfo(slotId)
 end
 
 -- }}}
+
+function MySlot:GetPetActionInfo(slotId)
+    local name, _, isToken, _, _, _, spellID = GetPetActionInfo(slotId)
+
+    local msg = _MySlot.Slot()
+    msg.id = slotId
+    msg.type = MYSLOT_SPELL
+
+    if isToken then
+        msg.strindex = name
+        msg.index = 0
+    elseif spellID then
+        msg.index = spellID
+    elseif not name then
+        msg.index = 0
+        msg.type = MYSLOT_EMPTY
+    else
+        return nil
+    end
+
+    return msg
+end
 
 -- {{{ GetBindingInfo
 -- {{{ Serialzie Key
@@ -308,6 +330,16 @@ function MySlot:Export(opt)
         end
     end
 
+    msg.petslot = {}
+    if not opt.ignorePetActionBar then
+        for i = 1, NUM_PET_ACTION_SLOTS, 1 do
+            local m = self:GetPetActionInfo(i)
+            if m then
+                msg.petslot[#msg.petslot + 1] = m
+            end
+        end
+    end
+
     local ct = msg:Serialize()
     local t = { MYSLOT_VER, 86, 04, 22, 0, 0, 0, 0 }
     MergeTable(t, StringToTable(ct))
@@ -432,6 +464,7 @@ function MySlot:FindMacro(macroInfo)
     -- Return index if found or nil
     return localMacro[name .. "_" .. body] or localMacro[body]
 end
+
 -- {{{ FindOrCreateMacro
 function MySlot:FindOrCreateMacro(macroInfo)
     if not macroInfo then
@@ -477,15 +510,12 @@ function MySlot:FindOrCreateMacro(macroInfo)
         return nil
     end
 end
-
 -- }}}
 
 function MySlot:RecoverData(msg, opt)
     -- {{{ Cache Spells
     --cache spells
     local spells = {}
-
-
     for i = 1, GetNumSpellTabs() do
         local tab, tabTex, offset, numSpells, isGuild, offSpecID = GetSpellTabInfo(i);
         offSpecID = (offSpecID ~= 0)
@@ -520,7 +550,6 @@ function MySlot:RecoverData(msg, opt)
             end
         end
     end
-
     -- }}}
 
 
@@ -536,11 +565,24 @@ function MySlot:RecoverData(msg, opt)
     end
     -- }}}
 
+    local slotBucket = {}
 
     -- {{{ Macro
-    -- cache macro
     local macro = {}
-    MySlot:Clear("MACRO", opt.clearOpt.ignoreMacros) -- will cause duplicate macro
+
+    for _, s in pairs(msg.slot or {}) do
+        local slotId = s.id
+        local slotType = _MySlot.Slot.SlotType[s.type]
+        local index = s.index
+
+        if slotType == MYSLOT_MACRO then
+            if macro[index] then
+                table.insert(macro[index], slotId)
+            else
+                macro[index] = {slotId}
+            end
+        end
+    end
 
     for _, m in pairs(msg.macro or {}) do
         local macroId = m.id
@@ -549,38 +591,36 @@ function MySlot:RecoverData(msg, opt)
         local name = m.name
         local body = m.body
 
-        if not opt.actionOpt.ignoreMacros["ACCOUNT"] and macroId <= MAX_ACCOUNT_MACROS  then
-            macro[macroId] = self:FindOrCreateMacro({
-                ["oldid"] = macroId,
-                ["name"] = name,
-                ["icon"] = icon,
-                ["body"] = body,
-            })
+        local info = {
+            ["oldid"] = macroId,
+            ["name"] = name,
+            ["icon"] = icon,
+            ["body"] = body,
+        }
+
+        local newid = nil
+
+        if (not opt.actionOpt.ignoreMacros["ACCOUNT"] and macroId <= MAX_ACCOUNT_MACROS)
+        or (not opt.actionOpt.ignoreMacros["CHARACTOR"] and macroId > MAX_ACCOUNT_MACROS)
+        then
+            newid = self:FindOrCreateMacro(info)
         end
 
-        if not opt.actionOpt.ignoreMacros["CHARACTOR"] and macroId > MAX_ACCOUNT_MACROS  then
-            macro[macroId] = self:FindOrCreateMacro({
-                ["oldid"] = macroId,
-                ["name"] = name,
-                ["icon"] = icon,
-                ["body"] = body,
-            })
+        if not newid then
+            newid = self:FindMacro(info)
         end
 
-        if not macro[macroId] then
-            macro[macroId] = self:FindMacro({
-                ["oldid"] = macroId,
-                ["name"] = name,
-                ["icon"] = icon,
-                ["body"] = body,
-            })
+        if newid then
+            for _, slotId in pairs(macro[macroId] or {}) do
+                PickupMacro(newid)
+                PlaceAction(slotId)
+            end
+        else
+            MySlot:Print(L["Ignore unknown macro [id=%s]"]:format(macroId))
         end
     end
     -- }}} Macro
 
-    MySlot:Clear("ACTION", opt.clearOpt.ignoreActionBars)
-
-    local slotBucket = {}
 
     for _, s in pairs(msg.slot or {}) do
         local slotId = s.id
@@ -624,16 +664,16 @@ function MySlot:RecoverData(msg, opt)
                         end
                     elseif slotType == MYSLOT_ITEM then
                         PickupItem(index)
-                    elseif slotType == MYSLOT_MACRO then
-                        local macroid = macro[index]
+                    -- elseif slotType == MYSLOT_MACRO then
+                    --     local macroid = macro[index]
 
-                        if not macroid then
-                            MySlot:Print(L["Ignore unknown macro [id=%s]"]:format(index))
-                        end
+                    --     if not macroid then
+                    --         MySlot:Print(L["Ignore unknown macro [id=%s]"]:format(index))
+                    --     end
 
-                        if curType ~= MYSLOT_MACRO or curIndex ~= macroid then
-                            PickupMacro(macroid)
-                        end
+                    --     if curType ~= MYSLOT_MACRO or curIndex ~= macroid then
+                    --         PickupMacro(macroid)
+                    --     end
                     elseif slotType == MYSLOT_SUMMONPET and strindex and strindex ~= curIndex then
                         C_PetJournal.PickupPet(strindex, false)
                         if not GetCursorInfo() then
@@ -662,9 +702,7 @@ function MySlot:RecoverData(msg, opt)
                     ClearCursor()
                 end
             end) then
-            MySlot:Print(L
-            ["[WARN] Ignore slot due to an unknown error DEBUG INFO = [S=%s T=%s I=%s] Please send Importing Text and DEBUG INFO to %s"]
-            :format(slotId, slotType, index, MYSLOT_AUTHOR))
+            MySlot:Print(L["[WARN] Ignore slot due to an unknown error DEBUG INFO = [S=%s T=%s I=%s] Please send Importing Text and DEBUG INFO to %s"]:format(slotId, slotType, index, MYSLOT_AUTHOR))
         end
     end
 
@@ -675,10 +713,6 @@ function MySlot:RecoverData(msg, opt)
                 ClearCursor()
             end
         end
-    end
-
-    if opt.clearOpt.ignoreBinding then
-        MySlot:Clear("BINDING")
     end
 
     if not opt.actionOpt.ignoreBinding then
@@ -710,6 +744,34 @@ function MySlot:RecoverData(msg, opt)
         SaveBindings(GetCurrentBindingSet())
     end
 
+
+    if not opt.actionOpt.ignorePetActionBar then
+        local pettoken = {}
+        for i = 1, NUM_PET_ACTION_SLOTS, 1 do
+            local name, _, isToken = GetPetActionInfo(i);
+            if isToken then
+                pettoken[name] = i
+            end
+        end
+
+        for _, p in pairs(msg.petslot or {}) do
+            if p.strindex then
+                local slot = pettoken[p.strindex]
+                if slot then
+                    PickupPetAction(slot)
+                    PickupPetAction(p.id)
+                end
+            elseif p.index then
+                PickupPetSpell(p.index)
+                PickupPetAction(p.id)
+            elseif p.type == MYSLOT_EMPTY then
+                PickupPetAction(p.id)
+            end
+            ClearCursor()
+        end
+    end
+
+
     MySlot:Print(L["All slots were restored"])
 end
 
@@ -727,7 +789,7 @@ function MySlot:Clear(what, opt)
                 DeleteMacro(i)
             end
         end
-    
+
         if opt["CHARACTOR"] then
             for i = MAX_ACCOUNT_MACROS + MAX_CHARACTER_MACROS,  MAX_ACCOUNT_MACROS + 1, -1 do
                 DeleteMacro(i)