+++ _wotlk/Interface/Addons/Myslot/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Myslot
 
-## [v5.16.0](https://github.com/tg123/myslot/tree/v5.16.0) (2024-06-29)
-[Full Changelog](https://github.com/tg123/myslot/commits/v5.16.0) 
+## [v5.17.1](https://github.com/tg123/myslot/tree/v5.17.1) (2024-06-30)
+[Full Changelog](https://github.com/tg123/myslot/commits/v5.17.1) 
 
-- feat: Update GUI label for "Other" in settings menu  
+- Merge pull request #54 from hewozuoai/hewozuoai/fix-action-bar-label  
+    Closes #53 - Fix ACTIONBAR\_LABEL text